requirements.txt
numpy
torch
pillow
notebook
matplotlib
chainercv
tqdm
pycocotools
